Added modularity
